bug: removed amplification of region for nanoseq masks
Previously, when generating the bed from the vcf, the defined bases
were added before and after the mutation. This was useful for lowcomplex
and lowmappability filters but not for nanoseq masks. Now, there
are two differentiated processes, one adding some bases and other
adding none. For this to work, the default number
of bases added is modified to 0.